[pam_bt_misc] Split code from header
