+++ pam/src/pam_bt_misc.h
@@ -1,14 +1,16 @@
 #ifndef PAM_BT_MISC_H
 #define PAM_BT_MISC_H
 
+#include <ctype.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
 #define BT_MAC_LEN 17
 #define BT_MAX_CONN 7 //Bluetooth Adapters can only connect up to 7 devices
 #define LOG 1
 
 const char *trusted_dir_path = "/etc/proxy_auth/";
-
-#include <ctype.h>
-
 
 /*
 * Return 1 iff the address is a valid Bluetooth Address
@@ -22,46 +24,7 @@ const char *trusted_dir_path = "/etc/proxy_auth/";
 * @param log_fp: the file handle for the log file
 * @return: 1 iff the address is a valid Bluetooth Address. Else return 0 (False)
 */
-int verify_bt_addr(char *address, FILE *log_fp) {
-    char *curr = address;
-    int len = 0;
-
-    int is_valid = 1;
-
-    if (!address) {
-        return 0;
-    }
-
-    if (strlen(address) <= BT_MAC_LEN) {   
-        while (curr && len < strlen(address)) {
-            int is_div3 = (len + 1) % 3 == 0;
-            
-            if (    len > BT_MAC_LEN ||                                 //exceeds the BT String Len
-                    (isxdigit(*curr) == 0 && !is_div3) ||               //not Alphanumeric when it should be
-                    (isxdigit(*curr) == 0 && is_div3 && *curr != ':')   //not a colon when it should be
-                ) {
-                is_valid = 0;
-                break;
-            }
-
-            len++;
-            curr++;
-        }
-    }
-    else {
-        is_valid = 0;
-    }
-       
-    if (is_valid && len == BT_MAC_LEN) {
-        return 1;
-    }
-    
-    if (log_fp) {
-        fprintf(log_fp, "%s is an invalid Bluetooth Address\n", address);
-    }
-
-    return 0;
-}
+int verify_bt_addr(char *address, FILE *log_fp);
 
 /*
 * Return 1 iff the given address is one of the trusted devices the user trusts
@@ -71,14 +34,7 @@ int verify_bt_addr(char *address, FILE *log_fp) {
 * @param num_of_devices: the number of devices the user trusts
 * @return: 1 iff the given address is one of the trusted devices the user trusts. Else return 0
 */
-int is_dev_trusted(FILE *log_fp, char *dev, char **trusted_devices, int num_of_devices) {
-    for (int i = 0; i < num_of_devices; i++) {
-        if (strcmp(dev, trusted_devices[i]) == 0) {
-            return 1;
-        }
-    }
-    return 0;
-}
+int is_dev_trusted(FILE *log_fp, char *dev, char **trusted_devices, int num_of_devices);
 
 /*
 * Free the list of bluetooth MAC addresses from memory
@@ -86,12 +42,5 @@ int is_dev_trusted(FILE *log_fp, char *dev, char **trusted_devices, int num_of_d
 * @param device_list: the list of bluetooth MAC addresses
 * @param num_of_devices: the cardinality of device array - represents the number of bluetooth devices stored in the array
 */
-void free_device_list(char **device_list, int num_of_devices) {
-    if (device_list) {
-        for (int i = 0; i < num_of_devices; i++) {
-            free(device_list[i]);
-        }
-        free(device_list);
-    }
-}
+void free_device_list(char **device_list, int num_of_devices);
 #endif
